+++ cmd/zard/main.go
@@ -20,13 +20,15 @@ import (
 	"github.com/cosmos/cosmos-sdk/server"
 	"github.com/cosmos/cosmos-sdk/store"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/genaccounts"
-	genaccscli "github.com/cosmos/cosmos-sdk/x/genaccounts/client/cli"
+	"github.com/cosmos/cosmos-sdk/x/auth"
 	genutilcli "github.com/cosmos/cosmos-sdk/x/genutil/client/cli"
 	"github.com/cosmos/cosmos-sdk/x/staking"
 )
 
+<<<<<<< HEAD:cmd/zard/main.go
 // zard custom flags
+=======
+>>>>>>> 3e3fae888d394d2d55ec7146984707aecdda4497:cmd/gaiad/main.go
 const flagInvCheckPeriod = "inv-check-period"
 
 var invCheckPeriod uint
@@ -49,14 +51,18 @@ func main() {
 	}
 
 	rootCmd.AddCommand(genutilcli.InitCmd(ctx, cdc, app.ModuleBasics, app.DefaultNodeHome))
-	rootCmd.AddCommand(genutilcli.CollectGenTxsCmd(ctx, cdc, genaccounts.AppModuleBasic{}, app.DefaultNodeHome))
+	rootCmd.AddCommand(genutilcli.CollectGenTxsCmd(ctx, cdc, auth.GenesisAccountIterator{}, app.DefaultNodeHome))
 	rootCmd.AddCommand(genutilcli.MigrateGenesisCmd(ctx, cdc))
-	rootCmd.AddCommand(genutilcli.GenTxCmd(ctx, cdc, app.ModuleBasics, staking.AppModuleBasic{},
-		genaccounts.AppModuleBasic{}, app.DefaultNodeHome, app.DefaultCLIHome))
+	rootCmd.AddCommand(
+		genutilcli.GenTxCmd(
+			ctx, cdc, app.ModuleBasics, staking.AppModuleBasic{},
+			auth.GenesisAccountIterator{}, app.DefaultNodeHome, app.DefaultCLIHome,
+		),
+	)
 	rootCmd.AddCommand(genutilcli.ValidateGenesisCmd(ctx, cdc, app.ModuleBasics))
-	rootCmd.AddCommand(genaccscli.AddGenesisAccountCmd(ctx, cdc, app.DefaultNodeHome, app.DefaultCLIHome))
+	rootCmd.AddCommand(AddGenesisAccountCmd(ctx, cdc, app.DefaultNodeHome, app.DefaultCLIHome))
 	rootCmd.AddCommand(client.NewCompletionCmd(rootCmd, true))
-	rootCmd.AddCommand(testnetCmd(ctx, cdc, app.ModuleBasics, genaccounts.AppModuleBasic{}))
+	rootCmd.AddCommand(testnetCmd(ctx, cdc, app.ModuleBasics, auth.GenesisAccountIterator{}))
 	rootCmd.AddCommand(replayCmd())
 
 	server.AddCommands(ctx, cdc, rootCmd, newApp, exportAppStateAndTMValidators)
@@ -72,11 +78,23 @@ func main() {
 }
 
 func newApp(logger log.Logger, db dbm.DB, traceStore io.Writer) abci.Application {
+<<<<<<< HEAD:cmd/zard/main.go
 	return app.NewZarApp(
+=======
+	var cache sdk.MultiStorePersistentCache
+
+	if viper.GetBool(server.FlagInterBlockCache) {
+		cache = store.NewCommitKVStoreCacheManager()
+	}
+
+	return app.NewGaiaApp(
+>>>>>>> 3e3fae888d394d2d55ec7146984707aecdda4497:cmd/gaiad/main.go
 		logger, db, traceStore, true, invCheckPeriod,
 		baseapp.SetPruning(store.NewPruningOptionsFromString(viper.GetString("pruning"))),
 		baseapp.SetMinGasPrices(viper.GetString(server.FlagMinGasPrices)),
-		baseapp.SetHaltHeight(uint64(viper.GetInt(server.FlagHaltHeight))),
+		baseapp.SetHaltHeight(viper.GetUint64(server.FlagHaltHeight)),
+		baseapp.SetHaltTime(viper.GetUint64(server.FlagHaltTime)),
+		baseapp.SetInterBlockCache(cache),
 	)
 }
 
@@ -85,6 +103,7 @@ func exportAppStateAndTMValidators(
 ) (json.RawMessage, []tmtypes.GenesisValidator, error) {
 
 	if height != -1 {
+<<<<<<< HEAD:cmd/zard/main.go
 		zApp := app.NewZarApp(logger, db, traceStore, false, uint(1))
 		err := zApp.LoadHeight(height)
 		if err != nil {
@@ -94,4 +113,16 @@ func exportAppStateAndTMValidators(
 	}
 	zApp := app.NewZarApp(logger, db, traceStore, true, uint(1))
 	return zApp.ExportAppStateAndValidators(forZeroHeight, jailWhiteList)
+=======
+		gapp := app.NewGaiaApp(logger, db, traceStore, false, uint(1))
+		err := gapp.LoadHeight(height)
+		if err != nil {
+			return nil, nil, err
+		}
+		return gapp.ExportAppStateAndValidators(forZeroHeight, jailWhiteList)
+	}
+
+	gapp := app.NewGaiaApp(logger, db, traceStore, true, uint(1))
+	return gapp.ExportAppStateAndValidators(forZeroHeight, jailWhiteList)
+>>>>>>> 3e3fae888d394d2d55ec7146984707aecdda4497:cmd/gaiad/main.go
 }