+++ CHANGELOG.md
@@ -26,7 +26,8 @@ Types of changes (Stanzas):
 "Improvements" for changes in existing functionality.
 "Deprecated" for soon-to-be removed features.
 "Bug Fixes" for any bug fixes.
-"Breaking" for breaking API changes.
+"Client Breaking" for breaking CLI commands and REST routes.
+"State Machine Breaking" for breaking the AppState
 
 Ref: https://keepachangelog.com/en/1.0.0/
 -->
@@ -34,6 +35,36 @@ Ref: https://keepachangelog.com/en/1.0.0/
 # Changelog
 
 ## [Unreleased]
+
+### Features
+
+* (gaiad) [\#119](https://github.com/cosmos/gaia/pull/119) Add support for the `--inter-block-cache` CLI
+flag and configuration.
+* (gaiacli) [\#132](https://github.com/cosmos/gaia/pull/132) Add `tx decode` command to decode
+Base64 encoded transactions.
+
+## [v2.0.1] - 2019-09-20
+
+### Features
+
+* (gaiad) [\#119](https://github.com/cosmos/gaia/pull/119) Add support for the `--halt-time` CLI flag and configuration.
+
+### Improvements
+
+* [\#119](https://github.com/cosmos/gaia/pull/119) Refactor and upgrade Circle CI
+configuration.
+* (sdk) Update SDK version to v0.37.1
+
+## [v2.0.0] - 2019-08-22
+
+### Bug Fixes
+
+* [\#104](https://github.com/cosmos/gaia/issues/104) Fix `ModuleAccountAddrs` to
+not rely on the `x/supply` keeper to get module account addresses for blacklisting.
+
+### State Machine Breaking Changes
+
+* (sdk) Update SDK version to v0.37.0
 
 ## [v1.0.0] - 2019-08-13
 
@@ -92,5 +123,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 <!-- Release links -->
 
-[Unreleased]: https://github.com/cosmos/gaia/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/cosmos/gaia/compare/v2.0.1...HEAD
+[v2.0.1]: https://github.com/cosmos/gaia/releases/tag/v2.0.1
+[v2.0.0]: https://github.com/cosmos/gaia/releases/tag/v2.0.0
 [v1.0.0]: https://github.com/cosmos/gaia/releases/tag/v1.0.0